feat: support fractional top-up amounts

+++ controller/topup_waffo_pancake_test.go
@@ -53,7 +53,7 @@ func TestGetWaffoPancakePayMoney(t *testing.T) {
 
 	testCases := []struct {
 		name             string
-		amount           int64
+		amount           float64
 		group            string
 		quotaDisplayType string
 		expected         float64
@@ -67,7 +67,7 @@ func TestGetWaffoPancakePayMoney(t *testing.T) {
 		},
 		{
 			name:             "tokens display converts quota to display units before pricing",
-			amount:           int64(common.QuotaPerUnit * 3),
+			amount:           common.QuotaPerUnit * 3,
 			group:            "vip",
 			quotaDisplayType: operation_setting.QuotaDisplayTypeTokens,
 			expected:         4.5,
