+++ setting/payment_waffo_pancake.go
@@ -9,7 +9,7 @@ var (
 	WaffoPancakePrivateKey string
 	WaffoPancakeReturnURL  string
 	WaffoPancakeUnitPrice  float64 = 1.0
-	WaffoPancakeMinTopUp   int     = 1
+	WaffoPancakeMinTopUp   float64 = 1
 	WaffoPancakeStoreID    string
 	WaffoPancakeProductID  string
 )
